Access Control

+++ app/Http/Controllers/PostsController.php
@@ -8,6 +8,17 @@ use App\Post;
 class PostsController extends Controller
 {
     /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+        
+    }
+
+    /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
@@ -69,6 +80,10 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+        // check for correct user
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error', 'Unauthorized Page');
+        }
         return view('posts.edit')->with('post' , $post);
     }
 
@@ -101,6 +116,11 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+        // check for correct user
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error', 'Unauthorized Page');
+        }
+        
         $post->delete();
         return redirect('/posts')->with('success', 'Post Removed');
     }
